Move buttons in header to their own components
